Refactored how CUCM objects are created
Previously, if any arguments were supplied to the constructor of a
BaseCUCMModel object, it would attempt to pass all of those arguments to
the 'get' AXL operation for the corresponding object. If an object was
not found, or if too many arguments were provided, the object would
fail to instantiate. Now, any number of arguments can be passed to the
constructor, and they will be handled in the following way:

Try:
Select and pass only valid arguments to the 'get' AXL operation.
If an AXL object is found, populate the CUCM object with the AXL object
attributes.

Except:
If no object is found, pass all arguments to the object's xtype
constructor, and populate the CUCM object with the xtype object
attributes. The object can then be added to Call Manager with the
create() method.